move types out of pocketbase directory
