+++ app/messages/new-gc.tsx
@@ -1,6 +1,6 @@
 import { pocketbase } from '@/features/pocketbase'
 import { useAppStore } from '@/features/stores'
-import { Profile } from '@/features/pocketbase/types'
+import { Profile } from '@/features/types'
 import { c, s } from '@/features/style'
 import { XStack, YStack } from '@/ui'
 import { Avatar } from '@/ui/atoms/Avatar'
