+++ features/home/feed.tsx
@@ -5,7 +5,7 @@ import { Dimensions, Pressable, ScrollView, View } from 'react-native'
 import { Ticker } from '@/features/home/Ticker'
 import { pocketbase } from '@/features/pocketbase'
 import { useAppStore } from '@/features/stores'
-import type { ExpandedItem } from '@/features/pocketbase/types'
+import type { ExpandedItem } from '@/features/types'
 import { c, s } from '@/features/style'
 import { DismissKeyboard, Heading, Text, XStack, YStack } from '@/ui'
 import SearchBottomSheet from '@/ui/actions/SearchBottomSheet'
