+++ features/home/search.tsx
@@ -3,7 +3,7 @@ import { s } from '../style'
 import { useEffect, useState } from 'react'
 import { pocketbase } from '@/features/pocketbase'
 import { useAppStore } from '@/features/stores'
-import { ExpandedItem, Profile } from '@/features/pocketbase/types'
+import { ExpandedItem, Profile } from '@/features/types'
 import UserListItem from '@/ui/atoms/UserListItem'
 import { router } from 'expo-router'
 import { Text, View } from 'react-native'
