+++ features/messaging/conversations-screen.tsx
@@ -5,7 +5,7 @@ import { pocketbase } from '@/features/pocketbase'
 import { useAppStore } from '@/features/stores'
 import SwipeableConversation from '@/ui/messaging/SwipeableConversation'
 import { router } from 'expo-router'
-import { Conversation } from '@/features/pocketbase/types'
+import { Conversation } from '@/features/types'
 import ConversationList from '@/ui/messaging/ConversationList'
 import { Ionicons } from '@expo/vector-icons'
 import { Link } from 'expo-router'
