+++ features/messaging/saves.tsx
@@ -9,7 +9,7 @@ import { Link, router } from 'expo-router'
 import { useEffect, useState } from 'react'
 import { Pressable, Text, TextInput, View, useWindowDimensions } from 'react-native'
 import Animated, { useSharedValue, useAnimatedStyle, withTiming } from 'react-native-reanimated'
-import { Conversation } from '@/features/pocketbase/types'
+import { Conversation } from '@/features/types'
 import SwipeableUser from '@/ui/atoms/SwipeableUser'
 import { t } from '@/features/style'
 import { useSafeAreaInsets } from 'react-native-safe-area-context'
