+++ features/types.ts
@@ -1,4 +1,4 @@
-import * as PBTypes from './pocketbase-types'
+import * as PBTypes from '@/features/pocketbase/pocketbase-types'
 
 export type CompleteRef = PBTypes.RefsRecord
 export type Profile = PBTypes.UsersRecord
