+++ ui/actions/SearchBottomSheet.tsx
@@ -9,7 +9,7 @@ import { Pressable, Text, TextInput, View } from 'react-native'
 import { Heading } from '../typo/Heading'
 import { XStack, YStack } from '../core/Stacks'
 import { Button } from '../buttons/Button'
-import { CompleteRef, Profile } from '@/features/pocketbase/types'
+import { CompleteRef, Profile } from '@/features/types'
 import { pocketbase } from '@/features/pocketbase'
 import { useAppStore } from '@/features/stores'
 import { SimplePinataImage } from '../images/SimplePinataImage'
