+++ ui/grid/DraggableGrid.tsx
@@ -2,7 +2,7 @@ import React, { useState, useRef } from 'react'
 import { ScrollView, TouchableOpacity, View, Text, FlatList, ScrollViewProps } from 'react-native'
 import { GridTile } from '../grid/GridTile'
 import { s } from '@/features/style'
-import { CompleteRef } from '@/features/pocketbase/types'
+import { CompleteRef } from '@/features/types'
 
 // @ts-ignore
 import { DraggableGrid as DraggableGridComponent } from 'react-native-draggable-grid'
