+++ ui/grid/GridTileWrapper.tsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react'
 import { YStack } from '../core/Stacks'
 import { TouchableOpacity, Pressable, Text } from 'react-native'
 import { base } from '@/features/style'
-import { GridTileType } from '@/features/pocketbase/types'
+import { GridTileType } from '@/features/types'
 import { useUIStore } from '../state'
 import Ionicons from '@expo/vector-icons/Ionicons'
 import { c } from '@/features/style'
