+++ ui/profiles/MyProfile.tsx
@@ -1,8 +1,8 @@
 import { pocketbase } from '@/features/pocketbase'
 import { useAppStore } from '@/features/stores'
 import { getBacklogItems, getProfileItems } from '@/features/stores/items'
-import type { ExpandedProfile } from '@/features/pocketbase/types'
-import { ExpandedItem } from '@/features/pocketbase/types'
+import type { ExpandedProfile } from '@/features/types'
+import { ExpandedItem } from '@/features/types'
 import { s } from '@/features/style'
 import BottomSheet from '@gorhom/bottom-sheet'
 import { useShareIntentContext } from 'expo-share-intent'
