+++ ui/profiles/SwipeableBacklogItem.tsx
@@ -3,7 +3,7 @@ import { Ionicons } from '@expo/vector-icons'
 import { Pressable, Linking, Text, View } from 'react-native'
 import { XStack, YStack } from '../core/Stacks'
 import { SimplePinataImage } from '../images/SimplePinataImage'
-import { ExpandedItem } from '@/features/pocketbase/types'
+import { ExpandedItem } from '@/features/types'
 import Swipeable from 'react-native-gesture-handler/ReanimatedSwipeable'
 import Reanimated, { SharedValue, useAnimatedStyle } from 'react-native-reanimated'
 import { useUIStore } from '../state'
