+++ ui/profiles/sheets/AddRefSheet.tsx
@@ -2,7 +2,7 @@ import { pocketbase } from '@/features/pocketbase'
 import { useAppStore } from '@/features/stores'
 import { getProfileItems } from '@/features/stores/items'
 import { RefsRecord } from '@/features/pocketbase/pocketbase-types'
-import { ExpandedItem, StagedItemFields } from '@/features/pocketbase/types'
+import { ExpandedItem, StagedItemFields } from '@/features/types'
 import { c, s } from '@/features/style'
 import { AddedNewRefConfirmation } from '@/ui/actions/AddedNewRefConfirmation'
 import { ChooseReplaceItemMethod } from '@/ui/actions/ChooseReplaceItemMethod'
