+++ ui/profiles/sorts.ts
@@ -1,4 +1,4 @@
-import type { Item } from '@/features/pocketbase/types'
+import type { Item } from '@/features/types'
 
 export const createdSort = (a: Item, b: Item) => {
   return new Date(a.created as string).valueOf() - new Date(b.created as string).valueOf()
